feat(lap-169): fix svg, create types and interfaces for questions
