+++ lib/types.ts
@@ -1,0 +1,29 @@
+export type AnswerModalProps = {
+  correctAnswer?: string | null;
+  explanation?: string | null;
+  onPressContinue: () => void;
+};
+
+export type MultipleChoiceQuestion = {
+  id: string;
+  paragraph: string;
+  question: string;
+  options: string[];
+  answer: number;
+};
+
+export type FillInTheBlankQuestion = {
+  id: string;
+  paragraph: string;
+  question: string;
+  answer: string;
+};
+
+export type MatchingQuestion = {
+  id: string;
+  question: string;
+  options: string[];
+  answer: number[];
+};
+
+export type Question = MultipleChoiceQuestion | FillInTheBlankQuestion | MatchingQuestion;
